fix: Add config for ToC title and use inline-block instead of float: left
float: left works fine on Chromium-based browsers, but there's a bug from Firefox 59 where opening and closing DevTools will cause the main content to be pushed down as if ToC is a block (abandoned to this day!). ref: Bug 1448995. As a temporary fix, set both the main content and ToC to be inline with each other, adjust .main's margin-left to accompany the ToC that goes before it