fix!: Zooming in may break the layout of content
`display: inline-block` creates whitespace where ToC was initially at, and .main's left margin is based off that. This causes the layout to break when zooming in i.e. .main may shift to the right to accompany the invisible ToC and `margin-left`. Fixed by setting the xl media query for nav + content + toc, and adjust .main accordingly to fit md, lg and xl.

As a side effect, if the device width cannot fit ToC, only nav and content will be shown and centered as with the current layout.

BREAKING CHANGES: Both nav.toc and .main are now set to `display: inline-block`. When making position changes to .main, also consider the width of the invisible block reserved for nav.toc (whose position is shifted).